ft(react): dispatch old workflow
