+++ .gitignore
@@ -1,9 +1,32 @@
-### Node ###
-# Logs
-logs
-*.log
+# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+/.pnp
+.pnp.js
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
 npm-debug.log*
 yarn-debug.log*
+yarn-error.log*
+yarn.lock
+
+
+ Logs
+logs
+*.log
 yarn-error.log*
 lerna-debug.log*
