+++ .travis.yml
@@ -8,12 +8,12 @@ cache:
   directories:
     - node_modules
 install:
-  - npm install
+  - yarn install
 before_script:
   - curl -L https://codeclimate.com/downloads/test-reporter/test-reporter-latest-linux-amd64 > ./cc-test-reporter
   - chmod +x ./cc-test-reporter
   - ./cc-test-reporter before-build
 script:
-  - npm test
+  -  yarn test:pre-commit
 after_success:
   - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
