--- fidio/.babelrc
@@ -1,4 +1,0 @@
-{
-  "presets": ["@babel/preset-env"],
-  "plugins": ["@babel/plugin-proposal-class-properties"]
-}
