--- fidio/README.md
@@ -1,53 +1,0 @@
-[![Build Status](https://travis-ci.org/Oluwasegun-AA/vidinfo.svg?branch=develop)](https://travis-ci.org/Oluwasegun-AA/vidinfo) [![Maintainability](https://api.codeclimate.com/v1/badges/ea36b527d5d4a998536f/maintainability)](https://codeclimate.com/github/Oluwasegun-AA/vidinfo/maintainability) <a href="https://codeclimate.com/github/Oluwasegun-AA/vidinfo/test_coverage"><img src="https://api.codeclimate.com/v1/badges/ea36b527d5d4a998536f/test_coverage" /></a>
-
-# make Gifs from Videos
-
-
-## Table of Contents
-
-* [Project Overview](#Project-Overview)
-* [Features](#Features)
-* [Demo](#demo)
-* [Built with](#built-with)
-* [Known Issues](#Known-issues)
-* [Installation](#Installation)
-* [Contributing](#contributing)
-* [License](#License)
-
-## Project Overview
-loading . . .
-
-## Features
-loading . . .
-
-## Demo
-loading . . .
-
-## Built with
-- `HTML 5`
-- `CSS`
-- `JavaScript`
-- `flow`
-- `Node.js`
-- `Electron`
-
-## Known issues
-loading . . .
-
-## Installation
-loading . . .
-
-## Contributing
->  Feel free to 🍴 fork this repository
-
->  👯 Clone this repository to your local machine using `{ loading . . . }
-
-> Make Contributions
-
-> 🔃 Create a new pull request using   { loading . . . }
-
-## License
-![GitHub](https://img.shields.io/github/license/mashape/apistatus.svg)
-
-- **[MIT license]()**
-- Copyright 2018 © <a href="https://twitter.com/Oluwasegun_AA" target="_blank">OLúwáségun</a>
