+++ src/styles/components/index.scss
