+++ src/tests/components/App.test.js
