+++ src/views/Test.jsx
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
-import logo from '../assets/logo.png';
+import img from '../assets/logo.png';
 
 class Test extends Component {
   render() {
     return (
       <div className="App">
         <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
+          <img src={img} className="App-logo" alt="logo" />
           <p>
             Edit <code>src/App.js</code> and save to reload.
           </p>
